Revert "IMS-VT: Change parameter type of data usage"
This reverts commit bbbab9c902f4df38551053c95ffa06772cd97696.

Conflicts:
	tests/src/com/android/server/telecom/testapps/TestVideoProvider.java

+++ tests/src/com/android/server/telecom/testapps/TestManagedVideoProvider.java
@@ -52,7 +52,7 @@ public class TestManagedVideoProvider extends Connection.VideoProvider {
         }
 
         @Override
-        public void onCallDataUsageChanged(RemoteConnection.VideoProvider rvp, long dataUsage) {
+        public void onCallDataUsageChanged(RemoteConnection.VideoProvider rvp, int dataUsage) {
             super.onCallDataUsageChanged(rvp, dataUsage);
         }
